Local custom Identity implementation on Sistrategia.Drive.Business Layer.
