+++ src/Sistrategia.Drive.Business/packages.config
@@ -1,0 +1,13 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
+  <package id="Microsoft.AspNet.Identity.Core" version="2.2.1" targetFramework="net45" />
+  <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.1" targetFramework="net45" />
+  <package id="Microsoft.AspNet.Identity.Owin" version="2.2.1" targetFramework="net45" />
+  <package id="Microsoft.Owin" version="2.1.0" targetFramework="net45" />
+  <package id="Microsoft.Owin.Security" version="2.1.0" targetFramework="net45" />
+  <package id="Microsoft.Owin.Security.Cookies" version="2.1.0" targetFramework="net45" />
+  <package id="Microsoft.Owin.Security.OAuth" version="2.1.0" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="4.5.11" targetFramework="net45" />
+  <package id="Owin" version="1.0" targetFramework="net45" />
+</packages>
