+++ src/Sistrategia.Drive.WebSite/App_Start/IdentityConfig.cs
@@ -1,85 +1,85 @@
-﻿using System;
-using System.Collections.Generic;
-//using System.Data.Entity;
-using System.Linq;
-using System.Security.Claims;
-using System.Threading.Tasks;
-using System.Web;
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Identity.EntityFramework;
-using Microsoft.AspNet.Identity.Owin;
-using Microsoft.Owin;
-using Microsoft.Owin.Security;
-using Sistrategia.Drive.WebSite.Models;
+﻿//using System;
+//using System.Collections.Generic;
+////using System.Data.Entity;
+//using System.Linq;
+//using System.Security.Claims;
+//using System.Threading.Tasks;
+//using System.Web;
+//using Microsoft.AspNet.Identity;
+//using Microsoft.AspNet.Identity.EntityFramework;
+//using Microsoft.AspNet.Identity.Owin;
+//using Microsoft.Owin;
+//using Microsoft.Owin.Security;
+//using Sistrategia.Drive.WebSite.Models;
 
-namespace Sistrategia.Drive.WebSite
-{
-    // Configure the application user manager used in this application. UserManager is defined in ASP.NET Identity and is used by the application.
-    public class ApplicationUserManager : UserManager<ApplicationUser>
-    {
-        public ApplicationUserManager(IUserStore<ApplicationUser> store)
-            : base(store) {
-        }
+//namespace Sistrategia.Drive.WebSite
+//{
+//    // Configure the application user manager used in this application. UserManager is defined in ASP.NET Identity and is used by the application.
+//    public class ApplicationUserManager : UserManager<ApplicationUser>
+//    {
+//        public ApplicationUserManager(IUserStore<ApplicationUser> store)
+//            : base(store) {
+//        }
 
-        public static ApplicationUserManager Create(IdentityFactoryOptions<ApplicationUserManager> options, IOwinContext context) {
-            var manager = new ApplicationUserManager(new UserStore<ApplicationUser>(context.Get<ApplicationDbContext>()));
+//        public static ApplicationUserManager Create(IdentityFactoryOptions<ApplicationUserManager> options, IOwinContext context) {
+//            var manager = new ApplicationUserManager(new UserStore<ApplicationUser>(context.Get<ApplicationDbContext>()));
 
-            // Configure validation logic for usernames
-            manager.UserValidator = new UserValidator<ApplicationUser>(manager) {
-                AllowOnlyAlphanumericUserNames = false,
-                RequireUniqueEmail = true
-            };
+//            // Configure validation logic for usernames
+//            manager.UserValidator = new UserValidator<ApplicationUser>(manager) {
+//                AllowOnlyAlphanumericUserNames = false,
+//                RequireUniqueEmail = true
+//            };
 
-            // Configure validation logic for passwords
-            manager.PasswordValidator = new PasswordValidator {
-                RequiredLength = 6,
-                RequireNonLetterOrDigit = false, //true,
-                RequireDigit = true,
-                RequireLowercase = true,
-                RequireUppercase = true,
-            };           
+//            // Configure validation logic for passwords
+//            manager.PasswordValidator = new PasswordValidator {
+//                RequiredLength = 6,
+//                RequireNonLetterOrDigit = false, //true,
+//                RequireDigit = true,
+//                RequireLowercase = true,
+//                RequireUppercase = true,
+//            };           
 
-            // Configure user lockout defaults
-            manager.UserLockoutEnabledByDefault = true;
-            manager.DefaultAccountLockoutTimeSpan = TimeSpan.FromMinutes(5);
-            manager.MaxFailedAccessAttemptsBeforeLockout = 5;
+//            // Configure user lockout defaults
+//            manager.UserLockoutEnabledByDefault = true;
+//            manager.DefaultAccountLockoutTimeSpan = TimeSpan.FromMinutes(5);
+//            manager.MaxFailedAccessAttemptsBeforeLockout = 5;
 
-            //// Register two factor authentication providers. This application uses Phone and Emails as a step of receiving a code for verifying the user
-            //// You can write your own provider and plug it in here.
-            //manager.RegisterTwoFactorProvider("Phone Code", new PhoneNumberTokenProvider<ApplicationUser> {
-            //    MessageFormat = "Your security code is {0}"
-            //});
-            //manager.RegisterTwoFactorProvider("Email Code", new EmailTokenProvider<ApplicationUser> {
-            //    Subject = "Security Code",
-            //    BodyFormat = "Your security code is {0}"
-            //});
-            //manager.EmailService = new EmailService();
-            //manager.SmsService = new SmsService();
+//            //// Register two factor authentication providers. This application uses Phone and Emails as a step of receiving a code for verifying the user
+//            //// You can write your own provider and plug it in here.
+//            //manager.RegisterTwoFactorProvider("Phone Code", new PhoneNumberTokenProvider<ApplicationUser> {
+//            //    MessageFormat = "Your security code is {0}"
+//            //});
+//            //manager.RegisterTwoFactorProvider("Email Code", new EmailTokenProvider<ApplicationUser> {
+//            //    Subject = "Security Code",
+//            //    BodyFormat = "Your security code is {0}"
+//            //});
+//            //manager.EmailService = new EmailService();
+//            //manager.SmsService = new SmsService();
 
-            var dataProtectionProvider = options.DataProtectionProvider;
-            if (dataProtectionProvider != null) {
-                manager.UserTokenProvider =
-                    new DataProtectorTokenProvider<ApplicationUser>(dataProtectionProvider.Create("ASP.NET Identity"));
-            }
+//            var dataProtectionProvider = options.DataProtectionProvider;
+//            if (dataProtectionProvider != null) {
+//                manager.UserTokenProvider =
+//                    new DataProtectorTokenProvider<ApplicationUser>(dataProtectionProvider.Create("ASP.NET Identity"));
+//            }
 
-            return manager;
-        }
-    }
+//            return manager;
+//        }
+//    }
 
 
-    // Configure the application sign-in manager which is used in this application.
-    public class ApplicationSignInManager : SignInManager<ApplicationUser, string>
-    {
-        public ApplicationSignInManager(ApplicationUserManager userManager, IAuthenticationManager authenticationManager)
-            : base(userManager, authenticationManager) {
-        }
+//    // Configure the application sign-in manager which is used in this application.
+//    public class ApplicationSignInManager : SignInManager<ApplicationUser, string>
+//    {
+//        public ApplicationSignInManager(ApplicationUserManager userManager, IAuthenticationManager authenticationManager)
+//            : base(userManager, authenticationManager) {
+//        }
 
-        public override Task<ClaimsIdentity> CreateUserIdentityAsync(ApplicationUser user) {
-            return user.GenerateUserIdentityAsync((ApplicationUserManager)UserManager);
-        }
+//        public override Task<ClaimsIdentity> CreateUserIdentityAsync(ApplicationUser user) {
+//            return user.GenerateUserIdentityAsync((ApplicationUserManager)UserManager);
+//        }
 
-        public static ApplicationSignInManager Create(IdentityFactoryOptions<ApplicationSignInManager> options, IOwinContext context) {
-            return new ApplicationSignInManager(context.GetUserManager<ApplicationUserManager>(), context.Authentication);
-        }
-    }
-}
+//        public static ApplicationSignInManager Create(IdentityFactoryOptions<ApplicationSignInManager> options, IOwinContext context) {
+//            return new ApplicationSignInManager(context.GetUserManager<ApplicationUserManager>(), context.Authentication);
+//        }
+//    }
+//}
