+++ src/Sistrategia.Drive.WebSite/App_Start/Startup.Auth.cs
@@ -5,7 +5,7 @@ using Microsoft.Owin;
 using Microsoft.Owin.Security.Cookies;
 //using Microsoft.Owin.Security.Google;
 using Owin;
-using Sistrategia.Drive.WebSite.Models;
+using Sistrategia.Drive.Business;
 
 namespace Sistrategia.Drive.WebSite
 {
@@ -15,8 +15,8 @@ namespace Sistrategia.Drive.WebSite
         public void ConfigureAuth(IAppBuilder app) {
             //// Configure the db context, user manager and signin manager to use a single instance per request
             app.CreatePerOwinContext(ApplicationDbContext.Create);
-            app.CreatePerOwinContext<ApplicationUserManager>(ApplicationUserManager.Create);
-            app.CreatePerOwinContext<ApplicationSignInManager>(ApplicationSignInManager.Create);
+            app.CreatePerOwinContext<SecurityUserManager>(SecurityUserManager.Create);
+            app.CreatePerOwinContext<SecuritySignInManager>(SecuritySignInManager.Create);
 
             // Enable the application to use a cookie to store information for the signed in user
             // and to use a cookie to temporarily store information about a user logging in with a third party login provider
@@ -25,7 +25,7 @@ namespace Sistrategia.Drive.WebSite
                 AuthenticationType = DefaultAuthenticationTypes.ApplicationCookie,
                 LoginPath = new PathString("/Account/Login"),
                 Provider = new CookieAuthenticationProvider {
-                    OnValidateIdentity = SecurityStampValidator.OnValidateIdentity<ApplicationUserManager, ApplicationUser>(
+                    OnValidateIdentity = SecurityStampValidator.OnValidateIdentity<SecurityUserManager, SecurityUser>(
                         validateInterval: TimeSpan.FromMinutes(30),
                         regenerateIdentity: (manager, user) => user.GenerateUserIdentityAsync(manager)
                     )
