+++ src/Sistrategia.Drive.WebSite/Sistrategia.Drive.WebSite.csproj
@@ -232,6 +232,12 @@
   <ItemGroup>
     <Folder Include="App_Data\" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Sistrategia.Drive.Business\Sistrategia.Drive.Business.csproj">
+      <Project>{51f4f067-2e3c-4ed9-af52-567491add227}</Project>
+      <Name>Sistrategia.Drive.Business</Name>
+    </ProjectReference>
+  </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
     <VSToolsPath Condition="'$(VSToolsPath)' == ''">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
